Moved texture compat functions to Util
